sheets_tools/sheets_requirements.txt
# Google Sheets Automation Dependencies
gspread>=5.7.0
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
pandas>=1.5.0
openpyxl>=3.0.0
